agents/requirements.txt
uv
ddgs
boto3
httpx
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents[otel]
strands-agents-tools
mcp
langfuse